Create day 5.1| password generator
I was particularly excited about this one as it not only challenged me and allowed me to learn more things independently, but it also relates to cyber security, which is one of my biggest interests within computer science! this challenge had two levels, and easy and hard version. this version is the harder version. were the password sequence was scrambled at the end making the password harder to hack and more durable.